Allow thumbnail images to be scaled.
This will allow us to set the thumbnail size to a decent viewport fraction
instead of relying on the effective thumbnail size (much like what we already
do with the main image).

For now, set it to a 1:1 effective pixel ratio, which fixes shitty scaling on
mobile browsers without having to rely on the "initial-scale" meta-tag.