Add viewport initial-scale=1 again to fix iOS/Safari
The brain-dead handling of the viewport and the "pixel-is-not-a-pixel" idiotic
movement in all mobile browsers is truly irking me, with Safari choosing a
different scale in portrait/landscape mode, causing the thumbnails to have a
larger scaling factor for absolutely no reason.

Not to mention that the viewport height still does *not* account for the
toolbar height correctly. Not going to fix that, dear Apple.