Provide a solution for no-javascript
This involves replacing the h2 in index.html with stuff that's shows the
thumbnails and links to the images, providing alt text and title
attributes based on the first line of the caption. It also uses some CSS
tricks to arrange the thumbnails in columns based on the width of the
display.

Hopefully this will help users linking to their gallery using social
media. At least Google+ seems to pick it up.