Replace the entire noscript section
Previously, the code only replaced the h2 element within the noscript
of the template file, but this did not work on subsequent runs. The h2
was gone and the replacement no longer worked. This commit replaces
the entire noscript tag with a new noscript tag containing the
overview we need, an so subsequent runs work as expected.