Took "todays" sitelen-mute fgallery Perl script https://github.com/kensanata/sitelen-mute/blob/d11ac46bcb6737b6d5c7e407bcc0549c0e335926/fgallery and reimplemented the patches to it and index.js which have been present in pkgsrc against fgallery version 1.8.2: https://github.com/NetBSD/pkgsrc-wip/tree/master/fgallery/patches
The galleries seem to work the same on NetBSD, so this might pave the
way to tracking a more active fgallery version in pkgsrc.

This is "work in progress":

The $ShareDir will need to be adjusted, Linuxisms to the cp command
will need to be handled better and a decision will need to be made
how the noscript.html stuff should work. I prefer keeping index.html
static and editing the noscript.html file.

Some testing would also be good ;-)